+++ source/Examples/Resources/SqlLogin/1-AddWindowsUser.ps1
@@ -1,7 +1,8 @@
 <#
     .DESCRIPTION
         This example shows how to ensure that the Windows user 'CONTOSO\WindowsUser',
-        Windows group 'CONTOSO\WindowsGroup', and the SQL Login 'SqlLogin' exists.
+        'CONTOSO\WindowsUser2', 'CONTOSO\WindowsUser3',
+        and Windows group 'CONTOSO\WindowsGroup' exists.
 #>
 
 Configuration Example
@@ -10,11 +11,7 @@ Configuration Example
     (
         [Parameter(Mandatory = $true)]
         [System.Management.Automation.PSCredential]
-        $SqlAdministratorCredential,
-
-        [Parameter(Mandatory = $true)]
-        [System.Management.Automation.PSCredential]
-        $LoginCredential
+        $SqlAdministratorCredential
     )
 
     Import-DscResource -ModuleName 'SqlServerDsc'
@@ -61,20 +58,6 @@ Configuration Example
             ServerName           = 'TestServer.company.local'
             InstanceName         = 'DSC'
             PsDscRunAsCredential = $SqlAdministratorCredential
-        }
-
-        SqlLogin 'Add_SqlLogin'
-        {
-            Ensure                         = 'Present'
-            Name                           = 'SqlLogin'
-            LoginType                      = 'SqlLogin'
-            ServerName                     = 'TestServer.company.local'
-            InstanceName                   = 'DSC'
-            LoginCredential                = $LoginCredential
-            LoginMustChangePassword        = $false
-            LoginPasswordExpirationEnabled = $true
-            LoginPasswordPolicyEnforced    = $true
-            PsDscRunAsCredential           = $SqlAdministratorCredential
         }
     }
 }